Add --filter flag
The normal approach to filtering would be to just use Unix pipes.
That strategy doesn't work well here, because the histogram max width is
determined by the line with the most commits. So if one file has been
changed 1000 times and all the rest are orders of magnitude smaller,
the bar chart will have a lot of unused space.